Create document with title + type

+++ space-sync-server/src/main/kotlin/com/ideaspace/workers/KafkaPartitionProcessor.kt
@@ -47,7 +47,7 @@ class KafkaPartitionProcessor(
             println("process record: ${record.value()}")
             if(messageCounter[key] == BATCH_LIMIT){
                     println("IO CONTEXT - Process flush kafka message to database")
-                    documentRepository.updateOffset(key, record.offset().toInt())
+                    documentRepository.updateOffset(key, record.offset())
                 messageCounter[key] = 0
             }
         }
